Finished 1st round of refactor. Added linting and babel. Removd defaults
* Added babel. Compiles using the CLI and `npm run compile`

* Finished initial refactor with es6 and cleaned up interfaces:
	- The functions don't take callbacks anymore, they return promises.
	- Default configurations is now simplified using es6 function defaults
	  and object destructuring syntax
	- http and tcp protocols now take the same arguments and return
	  the same format. The Promise resolves to an object representation of
	  the message response, and rejects to any errors that could have
	  happened along the way.

* Added linting with the Airbnb styleguide. Run with `npm run lint`
Keeps you honest!

* Removed REPL and CLI
This can be part of a longer discussion, but I don't think this
module should have the responsbility of being a CLI in addition to
implementing the core logic of the AP setup. I propose that any
desired CLI for this is built into `particle-cli`. That makes a much
more logical decision to me, as it continues to be a one-stop-shop for
all the things you may need to do with your Particle boards.

* Haven't ran tests yet.
But, I can `require` the package in, make some requests, and get the
errors rejecting that I expect to -- so something!

* Updated the license to a valid format in package.json